requirements.txt
streamlit
torch
torchvision
numpy
opencv-python
matplotlib
segmentation-models-pytorch
timm
albumentations
tqdm
Pillow
tensorboard
ultralytics